fix: preserve provider dispatch identity
